Some more in progress unit test junk
